build(api): Make test server exportable and move server starting to a separate file
